Added material domain randomization
Materials must be added to object material properties to be selected from.

Right now the method is very inefficient, i rebuild an entirely new list in order to move a random index to the top. for some reason blender only uses that material that is at the top of the list. I tried using things such as obj.active_material_select(), obj.data.update(), bpy.ops.object.editmode_toggle(), etc. the only way i found to successfully achieve this is to rebuild and reorder the list